changed the way we scrape the calendar
