--- server/calendar/src/scraping/mod.rs
@@ -1,4 +1,0 @@
-pub mod continous_scraping;
-mod main_api;
-mod task;
-pub mod tumonline_calendar;
